*Server - 함수 자료형 통일(컴파일러경고 제거) - cct->move() 후 컨트롤러 업데이트 테스트(해봤지만 안 되서 일단 주석) - move 패킷을 보낸 클라 자신도 패킷을 받아보면서 확인(주석)

+++ Minsik/AnimalKingdom_Server/AnimalKingdom_server/Net/Session.cpp
@@ -47,7 +47,7 @@ bool Session::onAccept(SOCKET socket, SOCKADDR_IN addrInfo)
 	return true;
 }
 
-void Session::onRecv(size_t recvSize)
+void Session::onRecv(UINT recvSize)
 {
 	// 패킷조립 및 실행
 	char *recvBuf = recvOver_.buffer_;
@@ -58,7 +58,7 @@ void Session::onRecv(size_t recvSize)
 		{
 			if (recvSize + storedBytes_ >= sizeof(HEADER))
 			{
-				int restHeaderSize = sizeof(HEADER) - storedBytes_;
+				UINT restHeaderSize = sizeof(HEADER) - storedBytes_;
 				memcpy(packetBuffer_ + storedBytes_, recvBuf, restHeaderSize);
 				recvBuf += restHeaderSize;
 				storedBytes_ += restHeaderSize; 
@@ -80,7 +80,7 @@ void Session::onRecv(size_t recvSize)
 			}
 		}
 
-		int restSize = totalBytes_ - storedBytes_;
+		UINT restSize = totalBytes_ - storedBytes_;
 
 		if (restSize <= recvSize)
 		{
